Poc_Api_Rest_Docker/django-server/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.7
django-extensions==3.2.3
django-timezone-field==6.0.1
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.6
kombu==5.3.4
packaging==23.2
pika==1.3.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.8
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13